+++ src/StrawberryShake/CodeGeneration/src/CodeGeneration.CSharp/Extensions/EnumBuilderExtensions.cs
@@ -1,0 +1,20 @@
+using System.Collections.Generic;
+using StrawberryShake.CodeGeneration.CSharp.Builders;
+
+namespace StrawberryShake.CodeGeneration.CSharp
+{
+    internal static class EnumBuilderExtensions
+    {
+        public static EnumBuilder AddElements(
+            this EnumBuilder builder,
+            IEnumerable<string> elements)
+        {
+            foreach (var element in elements)
+            {
+                builder.AddElement(element);
+            }
+
+            return builder;
+        }
+    }
+}
